[Tutorials] Simplify and rename make_tutorials.py to run_tutorials.y
This python script is no longer responsible for generating the tutorials docs, that is handled by `convert_ipynb_to_mdx.py`. We still need its functionality to run tutorials so let's just keep that part of the logic.
