+++ .github/workflows/deploy.yml
@@ -34,9 +34,12 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: "3.10"
-    - name: Convert Tutorials
+    - name: Run Tutorials
       run: |
         pip install -e ".[tutorial]"
+        python scripts/run_tutorials.py -w $(pwd)
+    - name: Convert Tutorials
+      run: |
         python3 scripts/convert_ipynb_to_mdx.py
     - name: Create new version in Docusaurus
       id: create-version
